refactor(ui): migrate Icon imports from common/ to @/components/ui
